+++ codeception.yml
@@ -1,3 +1,5 @@
+actor: Tester
+
 paths:
     tests: tests
     log: tests/_log
