+++ tests/.gitignore
@@ -1,2 +1,2 @@
-*SuiteGuy.php
+_helpers/*SuiteTester.php
 _helpers/_generated/*
