+++ tests/functional.suite.yml
@@ -1,6 +1,6 @@
 error_level: "E_ALL"
 
-class_name: FunctionalSuiteGuy
+class_name: FunctionalSuiteTester
 
 modules:
     enabled:
