+++ tests/functional/src/SimpleFormTest.php
@@ -2,40 +2,39 @@
 
 namespace Tests\Functional;
 
-use Codeception\TestCase\Test;
+use Codeception\Test\Unit;
 
 /**
  * @author Jáchym Toušek <enumag@gmail.com>
  */
-class SimpleFormTest extends Test
+class SimpleFormTest extends Unit
 {
-
     public function testRendering()
     {
-        $this->guy->amOnPage('/article/');
-        $this->guy->seeResponseCodeIs(200);
-        $this->guy->see(null, 'form');
-        $this->guy->see(null, 'input#form__token'); // CSRF protection
-        $this->guy->see(null, 'input[name="do"]'); // Signal for Nette/Application
-        $this->guy->see(null, 'button');
+        $this->tester->amOnPage('/article/');
+        $this->tester->seeResponseCodeIs(200);
+        $this->tester->see(null, 'form');
+        $this->tester->see(null, 'input#form__token'); // CSRF protection
+        $this->tester->see(null, 'input[name="do"]'); // Signal for Nette/Application
+        $this->tester->see(null, 'button');
     }
 
     public function testValidation()
     {
-        $this->guy->amOnPage('/article/');
-        $this->guy->seeResponseCodeIs(200);
-        $this->guy->fillField('#form_text', '');
-        $this->guy->click('button');
-        $this->guy->see('error', '.state');
-        $this->guy->see('This value should not be blank.');
+        $this->tester->amOnPage('/article/');
+        $this->tester->seeResponseCodeIs(200);
+        $this->tester->fillField('#form_text', '');
+        $this->tester->click('button');
+        $this->tester->see('error', '.state');
+        $this->tester->see('This value should not be blank.');
     }
 
     public function testSuccess()
     {
-        $this->guy->amOnPage('/article/');
-        $this->guy->seeResponseCodeIs(200);
-        $this->guy->fillField('#form_text', 'lorem ipsum');
-        $this->guy->click('button');
-        $this->guy->see('success', '.state');
+        $this->tester->amOnPage('/article/');
+        $this->tester->seeResponseCodeIs(200);
+        $this->tester->fillField('#form_text', 'lorem ipsum');
+        $this->tester->click('button');
+        $this->tester->see('success', '.state');
     }
 }
