+++ tests/integration.suite.yml
@@ -1,3 +1,3 @@
 error_level: "E_ALL"
 
-class_name: IntegrationSuiteGuy
+class_name: IntegrationSuiteTester
